requirements.txt
scikit-learn==1.1.3
numpy
pandas
matplotlib
streamlit==0.76.0
Pillow
seaborn